requirements.txt
polars
pathlib
pandas
pyarrow
openmeteo_requests
requests_cache
retry_requests
numpy
lightgbm
scikit-learn
matplotlib
shap
PyYAML
mlflow